move repository out of the 'canimod' GitHub organisation
The original reason for using an organisation was to ensure the project wasn't abandoned if the main contributor disappeared. However, turns out the organisation added an extra layer of maintenance and separated it from the SMAPI repository, while offering no benefit over simply using an open license and adding GitHub collaborators.